Consolidate the store initialization configuration files into one.
